backend/requirements.txt
Django>=5.0,<6
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-environ>=0.11
celery[redis]>=5.3
redis>=5.0
psycopg[binary]>=3.1
pgvector>=0.3
langchain>=0.3
langchain-core>=0.3
langchain-openai>=0.2
langchain-text-splitters>=0.3
langgraph>=0.2
pypdf>=4.0
python-docx>=1.1
tiktoken>=0.7
sentence-transformers>=3.0
# torch: installed in Dockerfile from https://download.pytorch.org/whl/cpu (CPU-only)
openai>=1.40
